Added PID controller and calibration file
